A start on ta jobs time visual table. mockData has been extracted from the mips dump and all excess fields for this table has been removed.
Attempting to use the TableVisual component to show a table with hortizontally grouped job status, and then group by discipline below.
Easy enough to get this grouped on discipline, showing count for WO's and Sum up the estimate hours.
But getting job status "on top of this" was not straight forward.
I dont think the Table visual will be able to do this as it stands now.
Also struggled to find any demos in docs for react-table or else where, that did this.